Use selection device wrapper.
