+++ src/reducers/indent.ts
@@ -2,6 +2,7 @@ import { alert, moveThought } from '../reducers'
 import { getNextRank, hasChild, rootedParentOf, prevSibling } from '../selectors'
 import { State } from '../@types'
 import { appendToPath, ellipsize, head, headRank, headValue, isEM, isRoot, parentOf, pathToContext } from '../util'
+import * as selection from '../device/selection'
 
 /** Increases the indentation level of the thought, i.e. Moves it to the end of its previous sibling. */
 const indent = (state: State) => {
@@ -31,7 +32,7 @@ const indent = (state: State) => {
   }
 
   // store selection offset before moveThought is dispatched
-  const offset = window.getSelection()?.focusOffset
+  const offset = selection.offset()
 
   const cursorNew = appendToPath(parentOf(cursor), prev, {
     ...head(cursor),
