+++ src/reducers/moveThoughtUp.ts
@@ -11,6 +11,7 @@ import {
   simplifyPath,
   getSortPreference,
 } from '../selectors'
+import * as selection from '../device/selection'
 
 /** Swaps the thought with its previous siblings. */
 const moveThoughtUp = (state: State) => {
@@ -59,7 +60,7 @@ const moveThoughtUp = (state: State) => {
   }
 
   // get selection offset before moveThought is dispatched
-  const offset = window.getSelection()?.focusOffset
+  const offset = selection.offset()
 
   const rankNew =
     prevThought && !isSorted
